+++ CHANGES.md
@@ -2,11 +2,15 @@
 
 Changes in descending order.
 
-### 09.08.2018 (v1.0.3-v1.0.5)
+### 09.08.2018 (v1.0.3-v1.0.6)
 
 <table>
+  <tr><td>example</td><td>improve handling for examples.</td></tr>
+  <tr><td>validate</td><td>action flag options should still run middleware validation.</td></tr>
   <tr><td>typings</td><td>typings were build by error with relative paths.</td></tr>
+  <tr><td>flag</td><td>resolve issue where aliases aren't honored.</td></tr>
   <tr><td>CHANGE.md</td><td>renamed to CHANGES to support default npm behavior.</td></tr>
+  <tr><td>extendFor</td><td>option type should not be overridden..</td></tr>
 </table>
 
 ### 08.26.2018 (v1.0.0-v1.0.2)